pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["Daniil-Solo <daniil.solo1723@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quarto-cli = "^1.4.553"
numpy = "^1.26.4"
matplotlib = "^3.8.4"
pandas = "^2.2.2"
seaborn = "^0.13.2"
folium = "^0.16.0"
python-dotenv = "^1.0.1"
pandas-stubs = "^2.2.1.240316"
types-seaborn = "^0.13.2.20240311"
snakemake = "^8.10.8"
scikit-learn = "^1.4.2"
click = "^8.1.7"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
mypy = "^1.9.0"
pre-commit = "^3.7.0"
jupyter = "^1.0.0"
nbqa = "^1.8.5"
kaggle = "^1.6.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "B", "D", "SIM", "I"]
ignore = ["F401", "D211", "D213"]

[tool.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
warn_unused_ignores = false
disable_error_code = "import-untyped"